refactor: mooved to helpers history folder
